+++ evolutionary_programming/objective_function/benchmarks.pxd
@@ -3,4 +3,4 @@ from .base_function cimport BaseFunction
 
 cdef class RastriginFunction(BaseFunction):
     cdef int _dimension
-    cpdef float evaluate(self, double[:] individual) noexcept nogil
+    cpdef double evaluate(self, double[:] individual) noexcept nogil
